move Windows DNS control out of `tun_windows.rs` too
