+++ rust/headless-client/src/interface/windows.rs
@@ -1,28 +1,34 @@
-use anyhow::Result;
+use anyhow::{Context as _, Result};
 use connlib_shared::{Cidrv4, Cidrv6};
-use ip_network::{IpNetwork, Ipv4Network, Ipv6Network};
 use std::{
-    collections::HashSet,
-    net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddrV4, SocketAddrV6},
-};
-use windows::Win32::NetworkManagement::IpHelper::{
-    CreateIpForwardEntry2, DeleteIpForwardEntry2, InitializeIpForwardEntry, MIB_IPFORWARD_ROW2,
+    net::{IpAddr, Ipv4Addr, Ipv6Addr},
+    os::windows::process::CommandExt,
+    process::{Command, Stdio},
 };
 
-pub(crate) struct InterfaceManager {
-    iface_idx: u32,
-    routes: HashSet<IpNetwork>,
-}
+// Hides Powershell's console on Windows
+// <https://stackoverflow.com/questions/59692146/is-it-possible-to-use-the-standard-library-to-spawn-a-process-without-showing-th#60958956>
+const CREATE_NO_WINDOW: u32 = 0x08000000;
+
+// wintun automatically appends " Tunnel" to this
+// TODO: De-dupe
+const TUNNEL_NAME: &str = "Firezone";
+
+pub(crate) struct InterfaceManager {}
 
 impl Drop for InterfaceManager {
     fn drop(&mut self) {
-        todo!()
+        if let Err(error) = connlib_shared::windows::dns::deactivate() {
+            tracing::error!(?error, "Failed to deactivate DNS control");
+        }
     }
 }
 
 impl InterfaceManager {
+    // Fallible on Linux
+    #[allow(clippy::unnecessary_wraps)]
     pub(crate) fn new() -> Result<Self> {
-        todo!()
+        Ok(Self {})
     }
 
     #[tracing::instrument(level = "trace", skip(self))]
@@ -32,8 +38,8 @@ impl InterfaceManager {
         ipv6: Ipv6Addr,
         dns_config: Vec<IpAddr>,
     ) -> Result<()> {
-        tracing::debug!("Setting our IPv4 = {}", config.ipv4);
-        tracing::debug!("Setting our IPv6 = {}", config.ipv6);
+        tracing::debug!("Setting our IPv4 = {}", ipv4);
+        tracing::debug!("Setting our IPv6 = {}", ipv6);
 
         // TODO: See if there's a good Win32 API for this
         // Using netsh directly instead of wintun's `set_network_addresses_tuple` because their code doesn't work for IPv6
@@ -45,7 +51,7 @@ impl InterfaceManager {
             .arg("address")
             .arg(format!("name=\"{TUNNEL_NAME}\""))
             .arg("source=static")
-            .arg(format!("address={}", config.ipv4))
+            .arg(format!("address={}", ipv4))
             .arg("mask=255.255.255.255")
             .stdout(Stdio::null())
             .status()?;
@@ -57,92 +63,17 @@ impl InterfaceManager {
             .arg("set")
             .arg("address")
             .arg(format!("interface=\"{TUNNEL_NAME}\""))
-            .arg(format!("address={}", config.ipv6))
+            .arg(format!("address={}", ipv6))
             .stdout(Stdio::null())
             .status()?;
 
-        connlib_shared::windows::dns::change(dns_config)
+        connlib_shared::windows::dns::change(&dns_config)
             .context("Should be able to control DNS")?;
         Ok(())
     }
 
     #[tracing::instrument(level = "trace", skip(self))]
-    pub(crate) async fn on_update_routes(
-        &mut self,
-        ipv4: Vec<Cidrv4>,
-        ipv6: Vec<Cidrv6>,
-    ) -> Result<()> {
-        let new_routes: HashSet<IpNetwork> = ipv4
-            .into_iter()
-            .map(|x| Into::<Ipv4Network>::into(x).into())
-            .chain(
-                ipv6.into_iter()
-                    .map(|x| Into::<Ipv6Network>::into(x).into()),
-            )
-            .collect();
-        if new_routes == self.routes {
-            return Ok(());
-        }
-
-        for new_route in new_routes.difference(&self.routes) {
-            self.add_route(*new_route)?;
-        }
-
-        for old_route in self.routes.difference(&new_routes) {
-            self.remove_route(*old_route)?;
-        }
-
-        // TODO: Might be calling this more often than it needs
-        connlib_shared::windows::dns::flush().expect("Should be able to flush Windows' DNS cache");
-        self.routes = new_routes;
-        Ok(())
-    }
-
-    // It's okay if this blocks until the route is added in the OS.
-    fn add_route(&self, route: IpNetwork) -> Result<()> {
-        const DUPLICATE_ERR: u32 = 0x80071392;
-        let entry = self.forward_entry(route);
-
-        // SAFETY: Windows shouldn't store the reference anywhere, it's just a way to pass lots of arguments at once. And no other thread sees this variable.
-        match unsafe { CreateIpForwardEntry2(&entry) }.ok() {
-            Ok(()) => Ok(()),
-            Err(e) if e.code().0 as u32 == DUPLICATE_ERR => {
-                tracing::debug!(%route, "Failed to add duplicate route, ignoring");
-                Ok(())
-            }
-            Err(e) => Err(e.into()),
-        }
-    }
-
-    // It's okay if this blocks until the route is removed in the OS.
-    fn remove_route(&self, route: IpNetwork) -> Result<()> {
-        let entry = self.forward_entry(route);
-
-        // SAFETY: Windows shouldn't store the reference anywhere, it's just a way to pass lots of arguments at once. And no other thread sees this variable.
-        unsafe { DeleteIpForwardEntry2(&entry) }.ok()?;
-        Ok(())
-    }
-
-    fn forward_entry(&self, route: IpNetwork) -> MIB_IPFORWARD_ROW2 {
-        let mut row = MIB_IPFORWARD_ROW2::default();
-        // SAFETY: Windows shouldn't store the reference anywhere, it's just setting defaults
-        unsafe { InitializeIpForwardEntry(&mut row) };
-
-        let prefix = &mut row.DestinationPrefix;
-        match route {
-            IpNetwork::V4(x) => {
-                prefix.PrefixLength = x.netmask();
-                prefix.Prefix.Ipv4 = SocketAddrV4::new(x.network_address(), 0).into();
-            }
-            IpNetwork::V6(x) => {
-                prefix.PrefixLength = x.netmask();
-                prefix.Prefix.Ipv6 = SocketAddrV6::new(x.network_address(), 0, 0, 0).into();
-            }
-        }
-
-        row.InterfaceIndex = self.iface_idx;
-        row.Metric = 0;
-
-        row
+    pub(crate) async fn on_update_routes(&mut self, _: Vec<Cidrv4>, _: Vec<Cidrv6>) -> Result<()> {
+        unimplemented!()
     }
 }
